pyproject.toml
[tool.poetry]
name = "spz course project"
version = "0.1.0"
description = ""
authors = ["Roman Matviitsiv <matviitsivroman@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pyusb = "^1.2.1"
black = "^22.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
